% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adl_bar_plots.R
\name{adl_bar_plots}
\alias{adl_bar_plots}
\title{Create bar plots in ADL's style}
\usage{
adl_bar_plots(
  df,
  x,
  y,
  group = NULL,
  fill = NULL,
  color = NULL,
  orientation = "vertical",
  col_text_size = 3.25,
  freq_plot = TRUE,
  tropes = FALSE,
  position = "stack",
  dodge_width = 0.8,
  dodge_reverse = TRUE,
  distance_from_bar = 0.25,
  ...
)
}
\arguments{
\item{df}{A dataframe or tibble.}
}
\description{
This function allows users to create different types of bar plots with
ADL's style. This a wrapper around \code{\link[ggplot2]{geom_col}},
\code{\link[ggplot2]{geom_text}}/\code{\link[ggplot2]{geom_label}}, and
\code{\link[ggplot2]{geom_errorbar}}. The purpose of this function is to save
the Center for Antisemitism Research's team time when making graphs.
}
