% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adl_corr_plot.R
\name{adl_corr_plot}
\alias{adl_corr_plot}
\title{Create bar plots in ADL's style}
\usage{
adl_corr_plot(df, x, y, color = NULL, fill = NULL, wt = NULL)
}
\arguments{
\item{df}{A dataframe or tibble. This can be piped in like with a normal
\code{\link[ggplot2]{ggplot}} function.}

\item{x}{Variable that goes in the x-axis. This is required.}

\item{y}{Variable that goes in the y-axis. This is required.}

\item{color}{Set the grouping variable for which the confidence intervals are
colored. This is used in stacked graphs and dodged graphs. If NULL, the
default, no grouping variable is used. Note: No need to set this if the
data is not grouped at all.}

\item{fill}{Set the grouping variable for which the inside of the bars are
colored. This is used in stacked graphs and dodged graphs. If NULL, the
default, no grouping variable is used. Note: No need to set this if the
data is not grouped at all.}
}
\description{
This function allows users to create different types of bar plots with
ADL's style. This a wrapper around \code{\link[ggplot2]{geom_col}},
\code{\link[ggplot2]{geom_text}}/,\code{\link[ggplot2]{geom_label}}, and
\code{\link[ggplot2]{geom_errorbar}}. This function was created to
standardize the graphs produced by CAR's team and to cut down on the amount
of time it takes to make these graphs.
}
\details{
As mentioned previously, this function is a wrapper around various \code{{ggplot2}}
functions in order to save time when making simple bar plots, dodged bar
plots, and stacked bar plots. As a result, it is not possible to combine
every element of the graph. If you would like to do so, we recommend using the
actual geoms from \code{{ggplot2}}.

Each of the elements included serve a purpose to allow you to customize the
graphs so that they look nice. Moreover, the arguments were created for the
types of graphs that CAR produces, namely mean plots and frequency plots.
}
