% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adl_palettes-documentation.R
\docType{data}
\name{adl_palettes}
\alias{adl_palettes}
\title{Complete list of available adl official color palettes}
\format{
A list of 9 \code{color} objects elements.
\describe{
\item{primary}{Primary categorical color}
\item{secondary}{Seconday categorical color}
\item{tertiary}{Tertiary categorical color}
\item{pid3}{Palette for partisanship/ideology (categorical)}
\item{categorical}{Palette for categorical data for up to 7 groups (categorical)}
\item{likert_6}{Palette for likert scales with 6 response options (diverging)}
\item{likert_4}{Palette for likert scales with 4 response options (diverging)}
\item{bluescale}{Blue scale color palette (sequential)}
\item{grayscale}{Grayscale color palette (sequential)}
}
}
\usage{
adl_palettes
}
\description{
A list of all color palettes to choose from, and preview directly in the console.
}
\details{
Pretty print and plot methods are powered by the \href{https://CRAN.R-project.org/package=prismatic}{prismatic}
package. Each palette is stored as a \code{color} object to enact this behaviour.

To view the data
}
\keyword{datasets}
