% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlgraphs-package.R
\docType{package}
\name{adlgraphs-package}
\alias{adlgraphs}
\alias{adlgraphs-package}
\title{adlgraphs: Create Graphs with ADL Style and Branding}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is designed to make it easier for folks at CAR at ADL to make graphs. The package contains all of the different themes related to various graph styles, a guide to which colors to use and when, and various functions created related to various graphs (mainly for coefficient plots).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JDenn0514/adlgraphs}
  \item \url{https://jdenn0514.github.io/adlgraphs/}
  \item Report bugs at \url{https://github.com/JDenn0514/adlgraphs/issues}
}

}
\author{
\strong{Maintainer}: Jacob Dennen \email{jdennen@adl.org}

}
\keyword{internal}
