% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_note.R
\name{attr_note}
\alias{attr_note}
\title{Get the note attribute}
\usage{
attr_note(x, data)
}
\arguments{
\item{x}{A vector object, the name of a column in a \code{data.frame}, or an
an actual \code{data.frame} object.}

\item{data}{A \code{data.frame} or \code{tibble} object. This should only be specified
when \code{x} is only the name of a column in a \code{data.frame}.}
}
\description{
This function makes it easy to get the note attribute from either an
individual vector or a data frame. NOTE: it is not possible to set or modify
the note attribute with this function.
}
