% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_question_preface.R
\name{attr_question_preface}
\alias{attr_question_preface}
\title{Get the question_preface attribute}
\usage{
attr_question_preface(x, data)
}
\arguments{
\item{x}{A vector object, the name of a column in a \code{data.frame}, or an
an actual \code{data.frame} object.}

\item{data}{A \code{data.frame} or \code{tibble} object. This should only be specified
when \code{x} is only the name of a column in a \code{data.frame}.}
}
\description{
This function makes it easy to get the question_preface from either an
individual vector or a data frame. NOTE: it is not possible to set or modify
the question preface attribute with this function.
}
