% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_survey_flow.R
\name{attr_survey_flow}
\alias{attr_survey_flow}
\title{Get the survey_flow attribute}
\usage{
attr_survey_flow(x, df)
}
\arguments{
\item{x}{A vector object, the name of a column in a \code{data.frame}, or an
an actual \code{data.frame} object.}

\item{df}{A \code{data.frame} or \code{tibble} object. This should only be specified
when \code{x} is only the name of a column in a \code{data.frame}.}
}
\description{
This function makes it easy to get the survey_flow from either an
individual vector or a data frame. NOTE: it is not possible to set or modify
the question preface attribute with this function.
}
