% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_transformation.R
\name{attr_transformation}
\alias{attr_transformation}
\title{Get the transformation attribute}
\usage{
attr_transformation(x, df)
}
\arguments{
\item{x}{A vector object, the name of a column in a \code{data.frame}, or an
an actual \code{data.frame} object.}

\item{df}{A \code{data.frame} or \code{tibble} object. This should only be specified
when \code{x} is only the name of a column in a \code{data.frame}.}
}
\description{
This function makes it easy to get the transformation attribute from either an
individual vector or a data frame. NOTE: it is not possible to set or modify
the transformation attribute with this function.
}
