% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{attributes}
\alias{attr_var_label}
\alias{attr_val_labels}
\alias{attr_levels}
\alias{attr_transformation}
\alias{attr_note}
\alias{attr_question_preface}
\alias{attr_survey_flow}
\title{These are a set of functions that make it easy to access attributes/metadata
from a vector or column in a \code{data.frame}. Note: these functions do not allow
you to change, set, or remove attributes.}
\usage{
attr_var_label(x, df)

attr_val_labels(x, df)

attr_levels(x, df)

attr_transformation(x, df)

attr_note(x, df)

attr_question_preface(x, df)

attr_survey_flow(x, df)
}
\arguments{
\item{x}{A vector object or the name of a column in a data.frame}

\item{df}{A \code{data.frame} or \code{tibble} object. This should be specified when \code{x}
is only the name of a column.}
}
\description{
These are a set of functions that make it easy to access attributes/metadata
from a vector or column in a \code{data.frame}. Note: these functions do not allow
you to change, set, or remove attributes.

Get the variable label attribute

Get the value labels attribute

Get the factor levels attribute from a vector

Get the transformation attribute from a vector

Get the note attribute

get question preface attribute
}
