% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polling_crosstabs.R
\name{calc_design_effect}
\alias{calc_design_effect}
\title{Calculate design effect (DEFF) for weighted data}
\usage{
calc_design_effect(weights)
}
\arguments{
\item{weights}{Vector of weights}
}
\value{
Design effect value
}
\description{
Calculate design effect (DEFF) for weighted data
}
