% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polling_crosstabs.R
\name{calc_moe}
\alias{calc_moe}
\title{Calculate margin of error for weighted proportion}
\usage{
calc_moe(p, n, deff)
}
\arguments{
\item{p}{Proportion (as percentage, e.g., 30 for 30\%)}

\item{n}{Unweighted sample size}

\item{deff}{Design effect}
}
\value{
Margin of error (as decimal, e.g., 0.0435 for 4.35 percentage points)
}
\description{
Calculate margin of error for weighted proportion
}
