% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polling_crosstabs.R
\name{calc_weighted_pct}
\alias{calc_weighted_pct}
\title{Calculate weighted percentage}
\usage{
calc_weighted_pct(values, weights)
}
\arguments{
\item{values}{Vector of values (typically a factor or character)}

\item{weights}{Vector of weights corresponding to values}
}
\value{
Named vector of weighted percentages
}
\description{
Calculate weighted percentage
}
