% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_demos.R
\name{clean_demos}
\alias{clean_demos}
\title{Clean up demographic variables}
\usage{
clean_demos(df)
}
\arguments{
\item{df}{A dataframe or tibble. This can be piped in like with a normal
\code{\link[ggplot2]{ggplot}} function.}
}
\value{
A data.frame object
}
\description{
This function makes it easier to clean demographic data but it really only
works when the data is set up in a very specific way. This is designed to
work with how ADL programs its surveys and to work in case any of the
variables listed below are not found in the data.
}
