% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create a data codebook}
\usage{
codebook(data)
}
\arguments{
\item{data}{An object of class \code{data.frame} or \code{tibble}}
}
\description{
This function takes a \code{data.frame} and creates a codebook. The new
object is a \code{data.frame} where each row contains different metadata info
found within each column from the original data set. Each column in the new
data set represents a different element of the underlying metadata. This is
similar to \code{labelled::look_for()} but shows more of the underlying metadata.
}
\details{
Currently, the new data set provided by this function has the following columns:
\itemize{
\item pos - The position of the variable.
\item variable - The name of the variable.
\item label - The variable label.
\item levels - If the variable is a factor, the factor levels are listed here.
\item value_labels - If the variable is a labelled vector, the value labels are
listed here.
\item transformation - An explanation of any potential data transformations the
variable underwent is listed here. This useful if you want to remember how
a variable was created without going back to the cleaning script.
\item question_preface - This contains the question preface. To elaborate, some
questions in surveys enable respondents to select multiple responses. Each
response gets it's own variable in the data. The value listed here is
supposed to contain the text that prefaced the response options. The
actual response option is listed under 'label'.
\item survey flow - This is used to indicate if there was an experiment or some
sort of branching involved in the survey flow.
\item note - A miscellaneous attribute in which you can add random information
about the variable that doesn't fit in the other attributes.
\item class - The class attribute of the variable.
\item type - The type of the variable.
\item missing - Indicates how many missing values there are.
\item range - If a numeric variable, shows the range of the values.
}
}
\examples{
# create the codebook
test_data_codebook <- codebook(test_data)
# view the codebook
test_data_codebook


}
