% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{dunnett}
\alias{dunnett}
\title{Run Dunnett's multiple comparisons test with one control.}
\usage{
dunnett(
  data,
  x,
  treats = NULL,
  group = NULL,
  wt = NULL,
  control = NULL,
  conf.level = 0.95,
  show_means = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{A numeric vector that will be used to calculate the means.}

\item{treats}{A variable whose values are used to determine if the means
are statistically significantly different from each other.}

\item{group}{A character or factor vector. It is a second level grouping
variable. If specified, the function checks to see if the values in
\code{treats} are statistically significant  within each level of this variable.}

\item{wt}{Weights. Add if you have a weighting variable and want to perform
Dunnett's test with weighted means,}

\item{control}{A string that specifies the level of the reference group
through which the others will be tested.}

\item{conf.level}{A number between 0 and 1 that signifies the width of the
desired confidence interval. Default is 0.95, which corresponds to a 95\%
confidence interval.}

\item{show_means}{Logical. Determines if the output should contain the mean
differences between the levels and the reference level or if it should show
the means of each level. Default is \code{FALSE}}
}
\description{
\code{dunnett()} calculates Dunnett's post hoc pairwise multiple comparisons
procedure. More simply, it calculates the mean of a variable (\code{x}) along the
different levels of a grouping variable (\code{treats}) and then determines if the
difference between the control/reference group and each level is statistically
significant.
}
\details{
While there are other functions that also perform Dunnett's Test, like
\code{\link[PMCMRplus:dunnettTest]{PMCMRplus::dunnettTest()}} and \code{\link[DescTools:DunnettTest]{DescTools::DunnettTest()}} to name a few,
there are a few key differences between this function and those. Firstly,
this function takes in a data frame or tibble. This was done so that it can
either be piped in or specified in the argument. See more in the example
section below.

Another important difference, is the addition of the \code{group} variable.
When \code{group} is specified, the function still performs the Dunnett test between
the variable in \code{x} and the variable in \code{treats}, but it does so along each
level of whatever variable is specified in \code{group}. You can see this in
action in the examples section below.

A third key difference is the output. This function outputs a tibble object
with a special class that enables it to be used with \code{prettytable()}.
Furthermore, there are two distinct outputs. The default output shows the
difference in means between each level and the control level and the
confidence intervals are based on this difference. However, if
\code{show_means = TRUE}, then the output shows the mean values for each level in
\code{treats} while still indicating the p-value.

Lastly, this function allows you to add weights to calculate weighted means.
}
\examples{
# Check to see if any of the education groups are significantly different
# from the control group (in this case "High School or Less") for conspiracy
# theory belief
dunnett(test_data, "acts_avg", "edu_f")
# now let's do the same but have it show the means
dunnett(test_data, "acts_avg", "edu_f", show_means = TRUE)

# now do the same as above but make "Graduate Degree" the control group
dunnett(test_data, "acts_avg", "edu_f", control = "Graduate Degree")
dunnett(test_data, "acts_avg", "edu_f", control = "Graduate Degree", show_means = TRUE)

# now let's add in partisanship (`pid_f3`) as the `group` variable. This let's us
# compare education levels within each level of `pid_f3`.
dunnett(test_data, "acts_avg", "edu_f", "pid_f3")
# and repeat that but showing the means
dunnett(test_data, "acts_avg", "edu_f", "pid_f3", show_means = TRUE)

}
