% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{dunnett_helper}
\alias{dunnett_helper}
\title{Perform Dunnett's test (mostly internal function)}
\usage{
dunnett_helper(
  data,
  x,
  treats,
  wt = NULL,
  control = NULL,
  conf.level = 0.95,
  show_means = FALSE,
  show_diffs = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{A numeric vector that will be used to calculate the means.
This can be a string or symbol.}

\item{treats}{A variable whose values are used to determine if the means
are statistically significantly different from each other. Should be
a factor or character vector. This can be a string or symbol.}

\item{wt}{Weights. Add if you have a weighting variable and want to perform
Dunnett's test with weighted means,}

\item{control}{A string that specifies the level of the reference group
through which the others will be tested.}

\item{conf.level}{A number between 0 and 1 that signifies the width of the
desired confidence interval. Default is 0.95, which corresponds to a 95\%
confidence interval.}

\item{show_means}{Logical. Determines if the output should contain the means
of each level. Default is \code{FALSE}}

\item{show_diffs}{Logical. Determines if the output should contain the
difference in means}
}
\description{
This function was created to mostly serve as an internal function for
\code{dunnett()}. It does all the same things as \code{dunnett()}, however,
you can't group the data and then do the calculation.
}
