% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{dunnett_pval_fun}
\alias{dunnett_pval_fun}
\title{Perform Dunnett's test (mostly internal function)}
\usage{
dunnett_pval_fun(
  data,
  x,
  treats,
  group = NULL,
  wt = NULL,
  n,
  control = NULL,
  conf.level = 0.95,
  show_means = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{A numeric vector that will be used to calculate the means.}

\item{treats}{A variable whose values are used to determine if the means
are statistically significantly different from each other.}

\item{group}{A character or factor vector. It is a second level grouping
variable. If specified, the function checks to see if the values in
\code{treats} are statistically significant  within each level of this variable.}

\item{wt}{Weights. Add if you have a weighting variable and want to perform
Dunnett's test with weighted means,}

\item{n}{If using a list of dataframes, this specifies which data.frame in
the list is being used to perform Dunnett's test.}

\item{control}{A string that specifies the level of the reference group
through which the others will be tested.}

\item{conf.level}{A number between 0 and 1 that signifies the width of the
desired confidence interval. Default is 0.95, which corresponds to a 95\%
confidence interval.}

\item{show_means}{Logical. Determines if the output should contain the mean
differences between the levels and the reference level or if it should show
the means of each level. Default is \code{FALSE}}
}
\description{
This function was created to mostly serve as an internal function for
\code{Dunnett}. It is the main working function that calculates the p-values.
}
