% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_dupe.R
\name{export_bot_dupe}
\alias{export_bot_dupe}
\title{Export data frame with only bots and duplicates}
\usage{
export_bot_dupe(data, filename, export_raw_data = TRUE, id = NULL)
}
\arguments{
\item{data}{A data.frame object you want to operate on}

\item{filename}{The name of the file you want to create}

\item{export_raw_data}{Logical. If TRUE, the default, exports the raw data. If FALSE,
exports only three columns: the ID, duplicate, and bot}

\item{id}{The unique ID variable. Only relevant when \code{export_raw_data} is FALSE.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
\code{export_bot_dupe()} is superseded because the function \code{export_bogus()}
does the same thing but also removes speedsters.

This function creates a file containing all responses suspected of being bots and/or
duplicates. Can create a .xlsx, .sav, or .csv file
}
