% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_val.R
\name{flip_val}
\alias{flip_val}
\title{Flip the valence of a vector}
\usage{
flip_val(x, label = NULL)
}
\arguments{
\item{x}{A vector of class \code{haven_labelled} or \code{numeric}}

\item{label}{A new variable label given to \code{x}.}
}
\value{
A numeric vector of the same length as \code{x}
}
\description{
Reverse the valence of a vector by keeping its values consistent, but
reversing its value labels and providing a new variable label. By reversing
the valence, I mean changing a negative statement to a positive one and
vice versa. Please check the vignette to have a better understanding of
exactly what this function does (still needs to be updated).
}
\examples{

# create a vector
blue <- sample(c(1:4), size = 20, replace = TRUE)
# add value labels
attr(blue, "labels") <- c(
  "Strongly agree" = 1, 
  "Somewhat agree" = 2, 
  "Somewhat disagree" = 3, 
  "Strongly disagree" = 4
)
# add a variable label
attr(blue, "label") <- "I like the color blue"

# make it a haven-labelled object (this is just for printing purposes)
class(blue) <- c("haven_labelled", "vctrs_vctr", "double")

# flip the valence
blue_flip <- flip_val(blue, "I dislike the color blue")
# show the output comapring the two
str(list(blue = blue, blue_flip = blue_flip))

# can also be used inside of dplyr::mutate()
library(dplyr)
  
new_df <- test_data \%>\% 
  dplyr::mutate(
    # flip the valence for deserving
    deserving_flip = flip_val(
      deserving, 
      "Groups at the bottom are not just as deserving as groups at the top"
    )
  ) \%>\% 
  # keep only the relevant columns
  dplyr::select(deserving, deserving_flip)

# show the output
str(new_df)



}
