% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{freq_fun}
\alias{freq_fun}
\title{Get the frequencies as a GT table}
\usage{
freq_fun(df, var, group1, group2, wt)
}
\arguments{
\item{df}{An object of type data.frame or tibble. If piping the df into the
function, this is not required.}

\item{var}{A character string. The variable with which want to get the
frequencies.}

\item{group1}{A character string. The first grouping variable.}

\item{group2}{A character string. The second grouping variable.}

\item{wt}{Weights. Add if you have a weighting variable and want to get
weighted frequencies}
}
\description{
This function creates a GT table of the frequencies of a specified variable
and has the ability to get the frequencies for up to two grouping variables.
While this can be used on it's own, it was created more to be used in
\code{get_all_freqs}, a function that outputs the frequencies of a set of
variables to a word doc.
}
