% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_freqs.R
\name{get_all_freqs}
\alias{get_all_freqs}
\title{Export frequencies for a set of variables to a word doc.}
\usage{
get_all_freqs(df, x, group, wt, show_genpop, file_name)
}
\arguments{
\item{df}{An object of type data.frame or tibble. If piping the df into the
function, this is not required.}

\item{x}{A vector of variables you want to get the frequencies for.}

\item{group}{A character string. The first grouping variable.}

\item{wt}{A character string. Add if you have a weighting variable and want
to get weighted frequencies}

\item{show_genpop}{Logical. Should there be a column showing the frequencies
for the general population}

\item{file_name}{A character string specifying the name of the file to be
created with the frequencies and where the file will be located. File must
end in .docx}
}
\description{
This function uses \code{\link[=get_freq_table]{get_freq_table()}} to get the frequencies for a set of variables
suppplied by the user. It then outputs these frequencies to a word doc.
}
