% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_var_labels.R
\name{get_all_var_labels}
\alias{get_all_var_labels}
\title{Get all variable labels}
\usage{
get_all_var_labels(df)
}
\arguments{
\item{df}{a data.frame or tibble object}
}
\description{
Create a vector containing character strings comprised of all the variable
labels for each column in a data.frame or tibble
}
