% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bogus_data.R
\name{get_bogus}
\alias{get_bogus}
\title{Get bogus data from a data frame}
\usage{
get_bogus(data, duration, cut_off = 0.3)
}
\arguments{
\item{data}{A data.frame object you want to operate on}

\item{duration}{The name of the time duration variable. Must be a string.}

\item{cut_off}{Specify what percentage of the median time should be used
to remove speedsters. Default is 0.3, which means people who's time to
complete is 0.3 that of the median completion time are removed.}
}
\description{
This function is the opposite of \code{remove_bogus()}. Instead of removing bots,
duplicates, and speedster from a data frame, it keeps them. This was designed to
make it easy to send to survey panel providers so they can remove them from
your data set.
}
