% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diffs.R
\name{get_diffs}
\alias{get_diffs}
\title{Calculate difference in means}
\usage{
get_diffs(
  data,
  x,
  treats,
  group = NULL,
  wt = NULL,
  ref_level,
  decimals = 3,
  conf.level = 0.95,
  na.rm = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{A numeric vector that will be used to calculate the means.
This can be a string or symbol.}

\item{treats}{A variable whose values are used to determine if the means
are statistically significantly different from each other. Should be
a factor or character vector. This can be a string or symbol.}

\item{group}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> A selection of columns to
group the data by in addition to \code{treats}. This operates very similarly
to \code{.by} from dplyr (for more info on that see \link[=dplyr_by]{?dplyr_by}).
See examples to see how it operates.}

\item{wt}{Weights. Add if you have a weighting variable and want to perform
Dunnett's test with weighted means.}

\item{ref_level}{A string that specifies the level of the reference group
through which the others will be tested.}

\item{decimals}{Number of decimals each number should be rounded to. Default is 2.}

\item{conf.level}{A number between 0 and 1 that signifies the width of the
desired confidence interval. Default is 0.95, which corresponds to a 95\%
confidence interval.}

\item{na.rm}{Logical. Determines if NAs should be removed}
}
\value{
A \code{data.frame} with columns for the grouping variables, the
treatment variable, the difference in means (\code{diffs}),
number of observations (\code{n}), low CI (\code{conf.low}), high CI
(\code{conf.high}), and p-value (\code{p_value}).
}
\description{
This function calculates the difference in means using a
bivariate regression, as well the p-value indicating
how significant each difference is. The main function doing the
calculations \code{lm()}.

NOTE: This function does not perform an actual Dunnet Test as it
does not calculate the quantile of the multivariate t-distribution
when determining the confidence intervals and p-values. If you need
to perform an actual Dunnett Test use the \code{dunnett()} function
instead. Please be aware that that function is far slower when
there are many comparison groups due to the nature of
\code{mvtnorm::qmvt()} and high dimensional data.
}
\examples{
# load dply for the pipe: \%>\% 
library(dplyr)
library(adlgraphs)

# Check to see if any of the partisan groups are significantly different
# from the control group (in this case "Democrat") for conspiracy
# theory belief
get_diffs(test_data, "acts_avg", "pid_f3")

# now do the same as above but make "Independent" the control group
get_diffs(test_data, "acts_avg", "pid_f3", ref_level = "Independent")

# now let's add in education (`edu_f2`) as the `group` variable. This let's us
# compare education levels within each level of `edu_f2`. Note how the arguments
# don't have to be strings
get_diffs(test_data, acts_avg, pid_f3, edu_f2)

# we can also group by multiple variables. Due to a small n, I'm going to use 
# `edu_f2` instead of `edu_f`. 
test_data \%>\% 
  dplyr::mutate(values_f2 = make_dicho(values)) \%>\% 
  get_diffs(acts_avg, treats = pid_f3, group = c(edu_f2, values_f2))

# now let's do those previous two calculations but using `dplyr::group_by()`
test_data \%>\% 
  dplyr::group_by(pid_f3) \%>\% 
  get_diffs(acts_avg, edu_f)

# we can also group by multiple
test_data \%>\% 
  dplyr::mutate(values_f2 = make_dicho(values)) \%>\% 
  dplyr::group_by(pid_f3, values_f2) \%>\% 
  get_diffs(acts_avg, edu_f2)

# If we want to show means and differences, set show_means to TRUE
# we don't need to set show_diffs to TRUE since that is the default
get_diffs(test_data, acts_avg, edu_f, show_means = TRUE)

# if we want to show means without differences, set `show_diffs = FALSE`
get_diffs(test_data, acts_avg, edu_f, show_means = TRUE, show_diffs = FALSE)


}
