% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polling_crosstabs.R
\name{get_var_label}
\alias{get_var_label}
\title{Get variable label or use variable name if no label exists}
\usage{
get_var_label(data, var_name)
}
\arguments{
\item{data}{Data frame}

\item{var_name}{Variable name}
}
\value{
Variable label or name
}
\description{
Get variable label or use variable name if no label exists
}
