% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_themes.R
\name{hc_theme_default}
\alias{hc_theme_default}
\title{The default theme for hc charts}
\usage{
hc_theme_default(
  hc,
  base_size = 16,
  legend_position = "right",
  axis_text = TRUE,
  axis_text_x = TRUE,
  axis_text_y = TRUE,
  grid = TRUE,
  grid_x_only = FALSE,
  grid_y_only = FALSE,
  ...
)
}
\arguments{
\item{hc}{A highcharter object}

\item{base_size}{Base font size, given in px Also controls the spacing in
the graph.}

\item{legend_position}{The position of the legend. Options are: "right",
"top", "bottom", or "none". "none" removes the legend. "right" is the
default.}

\item{axis_text}{Logical. Determines if BOTH axes should have labels. If
\code{TRUE}, the default, both axes are labelled. If \code{FALSE}, neither axis is
labelled. Note, this controls both axes. If you want to remove only one
axis, use the \code{axis_text_x} or \code{axis_text_y}.}

\item{axis_text_x}{Logical. Determines if the x-axis has labels. If \code{TRUE},
the default, the x-axis labels are shown. If \code{FALSE}, the x-axis labels are
removed from the plot.}

\item{axis_text_y}{Logical. Determines if the y-axis has labels. If \code{TRUE},
the default, the y-axis labels are shown. If \code{FALSE}, the y-axis labels are
removed from the plot.}

\item{grid}{Logical. Determines if ALL grid lines should appear. If \code{TRUE},
the default, all grid lines appear. If \code{FALSE}, all grid lines disappear,
regardless of what \code{grid_x_only} or \code{grid_y_only} are set to.}

\item{grid_x_only}{Logical. Determines if only x-axis grid lines (vertical
lines) should appear. If \code{FALSE}, the default, all grid lines appear. If
\code{TRUE}, only the x-axis grid lines appear and the y-axis grid lines will
disappear.}

\item{grid_y_only}{Logical. Determines if only y-axis grid lines (horizontal
lines) should appear. If \code{FALSE}, the default, all grid lines appear. If
\code{TRUE}, only the y-axis grid lines appear and the  x-axis grid lines will
disappear.}

\item{...}{Additional arguments passed}
}
\description{
This function creates the default theme that all ADL highcharter themes
are built off of. It's ggplot2 equivalent is \code{theme_default}.
}
\details{
Due to certain limitations of Highcharter, this is not quite as extensive as
\code{theme_default}. For example, there are no minor grid lines right now. If I
am able to figure it out then I might add them in. In addition, the faceting
arguments from \code{theme_default} are not present here
}
