% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_percent.R
\name{make_percent}
\alias{make_percent}
\title{Convert a numeric vector to a percent}
\usage{
make_percent(x, digits = 2, scale = 100)
}
\arguments{
\item{x}{A numeric vector you want to convert to percentages}

\item{digits}{How many decimals should the value be rounded to. Default is
2 which means it will show two decimal places, or the hundredth decimal.}

\item{scale}{A scaling factor: \code{x} will be multiplied by \code{scale} before
formatting. This is useful if the underlying data is very small or very
large. Default is 100.}
}
\description{
Convert a numeric vector to a percent
}
\examples{
# here's the default scale of 100
x <- c(0.0163, 0.95, 0.0008, 0.002)
make_percent(x)

# if the values have already been multiplied by 100 and you don't need to
# transform them, then make the scale 1
x <- c(1.63, 95, 0.08, 0.2)
make_percent(x, scale = 1)

# And if we want to round to the closest whole number set the digits to 0
x <- c(0.0163, 0.95, 0.0008, 0.002)
make_percent(x, digits = 0)


}
