% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_quarts.R
\name{make_quarts}
\alias{make_quarts}
\title{Create quartiles}
\usage{
make_quarts(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A factor vector with four levels the same length as \code{x}.
}
\description{
This is a simple function that takes a numeric vector as the input and returns
a factor vector with labels indicating which quartile it is in.
}
