% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_rev.R
\name{num_rev}
\alias{num_rev}
\title{Reverse a numeric function}
\usage{
num_rev(x)
}
\arguments{
\item{x}{A vector of class \code{haven_labelled} or \code{numeric}}
}
\value{
A numeric vector of the same length as \code{x}
}
\description{
Reverse a numeric variable while maintaining variable and value labels
if available. Also adds an attribute describing the transformation the
original variable underwent. Please check the vignette to have a better
understanding of exactly what this function does.
}
\examples{

library(dplyr)

test_data \%>\%
  # reverse the variable accept_isr
  mutate(accept_isr_rev = num_rev(accept_isr)) \%>\%
  select(starts_with("accept"))

}
