% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_conv.R
\name{pct_conv}
\alias{pct_conv}
\title{Make proper percent labels}
\usage{
pct_conv(data, var = pct)
}
\arguments{
\item{data}{A data frame or vector. Can be left blank if used during piping}

\item{var}{a variable we want to convert to a percentage. The value is \code{pct}
by default and should always be pct.}
}
\description{
This function is a wrapper around a very common data transformation that is
done every time we make a frequency plot
}
