% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bogus_data.R
\name{remove_bogus}
\alias{remove_bogus}
\title{Remove bots, duplicates, and/or speedsters from a data frame}
\usage{
remove_bogus(data, duration, cut_off = 0.3)
}
\arguments{
\item{data}{A data frame.}

\item{duration}{The name of the time duration variable. Must be a string.}

\item{cut_off}{Specify what percentage of the median time should be used
to remove speedsters. Default is 0.3, which means people who's time to
complete is 0.3 that of the median completion time are removed.}
}
\description{
This function removes respondents suspected of being bots or duplicate
survey takers, as well as previews, people under 18, and speedsters
from a data frame. It is made specifically for surveys programmed with
Qualtrics.
}
