% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_dupe.R
\name{remove_bot_dupe}
\alias{remove_bot_dupe}
\title{Remove bots and/or duplicates from a data frame}
\usage{
remove_bot_dupe(data)
}
\arguments{
\item{data}{A data.frame object you want to operate on}
}
\description{
This function removes respondents suspected of being bots or duplicate
survey takers, as well as previews and people under 18 from a data frame.
It is made specifically for surveys programmed with Qualtrics.
}
