% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_means.R
\name{row_means}
\alias{row_means}
\title{Calculate row means}
\usage{
row_means(cols, label = NULL, na.rm = TRUE)
}
\arguments{
\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> The variables you want
to use when calculating row means}

\item{label}{A string specifying the variable label. If not specified,
defaults to NULL}

\item{na.rm}{Determines if NAs should be removed. Default is TRUE.}
}
\description{
This function makes it easy to calculate row means for multiple variables.
It uses <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> syntax to determine which
variables to include in the operation.
}
\details{
This function also has the option of adding a new variable label attribute.
Furthermore, it automatically adds two more attributes: \code{transformation}
and \code{variables}. The \code{trasnformation} attribute basically explains how the
variable was created by saying "Took the average of..." and then lists
the variables included. \code{variables} just lists the variables included
in the operation.
}
\examples{
# load the dplyr package
library(dplyr)
# make a new df with the new column
new <- test_data \%>\%
  mutate(
    sdo_avg_new = row_means(
      # specify the variables involved in the row means
      cols = c(top_rev:deserving_flip),
      # specify the variable label
      label = "Social Dominance Orientation Average",
      # remove NAs
      na.rm = TRUE
    )
  )

# Show that the attributes
attributes(new$sdo_avg_new)

# show the output
new$sdo_avg_new

}
