% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_adl.R
\name{scale_adl}
\alias{scale_adl}
\title{ADL color scale}
\usage{
scale_adl(
  type = "categorical",
  palette = "base",
  aesthetic = c("fill", "color"),
  n,
  direction = "original",
  legend_title = NULL,
  legend_order = "original",
  ...
)
}
\arguments{
\item{type}{}

\item{legend_order}{}
}
\description{
Create sequential, diverging, and categorical color scales using official ADL
colors.
}
\details{
The function `scale_adl()` was created for survey data.
}
