% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_quant.R
\name{stat_density_quant}
\alias{stat_density_quant}
\title{Stat for density ridgeline plots}
\usage{
stat_density_quant(
  mapping = NULL,
  data = NULL,
  geom = geom,
  position = "stack",
  ...,
  bw = "sj",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  na.rm = FALSE,
  bounds = c(-Inf, Inf),
  show.legend = NA,
  inherit.aes = TRUE,
  quantile_lines = FALSE,
  calc_ecdf = FALSE,
  quantiles = 4
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no
plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{ggplot2::fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data. \code{\link[ggplot2:geom_density]{ggplot2::geom_density}}
is the default.}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{ggplot2::layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{bw}{The smoothing bandwidth to be used.
If numeric, the standard deviation of the smoothing kernel.
If character, a rule to choose the bandwidth, as listed in
\code{\link[stats:bandwidth]{stats::bw.nrd()}}. Note that automatic calculation of the bandwidth does
not take weights into account. Default is \code{sj}.}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible
to adjust the bandwidth while still using the a bandwidth estimator.
For example, \code{adjust = 1/2} means use half of the default bandwidth.}

\item{kernel}{Kernel. See list of available kernels in \code{\link[=density]{density()}}.}

\item{n}{number of equally spaced points at which the density is to be
estimated, should be a power of two, see \code{\link[=density]{density()}} for
details}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{bounds}{Known lower and upper bounds for estimated data. Default
\code{c(-Inf, Inf)} means that there are no (finite) bounds. If any bound is
finite, boundary effect of default density estimation will be corrected by
reflecting tails outside \code{bounds} around their closest edge. Data points
outside of bounds are removed with a warning.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{ggplot2::borders()}}.}

\item{quantile_lines}{Logical. Determines if quantile lines should be drawn
or not. FALSE is default.}

\item{calc_ecdf}{If \code{TRUE}, \code{stat_density_ridges} calculates an empirical
cumulative distribution function (ecdf) and returns a variable \code{ecdf} and
a variable \code{quantile}. Both can be mapped onto aesthetics via \code{stat(ecdf)}
and \code{stat(quantile)}, respectively.}

\item{quantiles}{Sets the number of quantiles the data should be broken into.
Used if either \code{calc_ecdf = TRUE} or \code{quantile_lines = TRUE}. If
\code{quantiles} is an integer then the data will be cut into that many equal
quantiles. If it is a vector of probabilities then the data will cut by them.}
}
\description{
This stat is the default stat that will be used in \code{geom_density_quant} when
I get around to making it. Nevertheless, it still works with
\code{\link[ggplot2]{geom_density}}. It is very similar to
\code{\link[ggplot2]{stat_density}} and
\code{\link[ggridges]{stat_density_ridges}} as it was built as a sort of
combination of the two. One of the key differences between this function and
those two is that this one uses the Sheather & Jones ("sj") as the default bandwidth
selector. This is done because this is a better bandwidth selector than
Silverman's ("nrd0") which is the default for the other two functions. In
addition, this function allows you to add quantile lines similar to
\code{\link[ggridges]{stat_density_ridges}}.
}
