% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_documentation.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{{dataset_title},}
\format{
A data frame with {nrow(df)} rows and {ncol(df)} variables:
\describe{
\item{resp_id}{Unique ID for each Respondent}
\item{wts}{Survey weighting variable}
\item{edu}{What is the highest level of school you have completed or the highest degree you have received?}
\item{pid_f3}{Political Partisanship}
\item{pol_part_rally}{Attended a political rally, speech or campaign event}
\item{pol_part_worked}{Worked or volunteered for a political party, candidate or campaign}
\item{pol_part_contact}{Contacted an elected official}
\item{pol_part_money}{Contributed money to a candidate running for public office or to a group working to elect a candidate}
\item{pol_part_social}{Publicly expressed your support for a political candidate on Facebook, Twitter/X, Instagram, or other social media}
\item{pol_part_attended}{Attended government meetings in your community, such as city or town council meetings}
\item{pol_part_none}{None of the above}
\item{top}{An ideal society requires some groups to be on top and others to be on the bottom}
\item{inferior}{Some groups of people are simply inferior to other groups}
\item{dominate}{No one group should dominate in society}
\item{deserving}{Groups at the bottom are just as deserving as groups at the top}
\item{special}{It is unfair for some groups in society to receive special treatment from the government}
\item{harder}{I have a harder time succeeding than my parents did}
\item{controlled}{Much of our lives are being controlled by plots hatched in secrecy}
\item{small}{Even though we live in a democracy, a few people will always run things anyway}
\item{run}{The people who really "run" the country are not known to the voters}
\item{big_events}{Big events like wars, recessions, and the outcomes of elections are controlled by small groups of people who are working in secret against the rest of us}
\item{accept_hamas}{If you had a close family member who supported Hamas, would your family and friends find it…}
\item{accept_isr}{If you had a close family member who supported Israel, would your family and friends find it…}
\item{dislike_jews}{How many of your close friends/family dislike Jews}
\item{stick_together}{Jews stick together more than other Americans}
\item{values}{Jews share my values.}
\item{head}{Jews always like to be at the head of things.}
\item{loyal}{Jews are more loyal to Israel than to America.}
\item{business_power}{Jews have too much power in the business world.}
\item{wall_street}{Jews have too much control and influence on Wall Street.}
\item{trad_n}{ADL Index}
\item{edu_f}{What is the highest level of school you have completed or the highest degree you have received?}
\item{edu_f2}{What is the highest level of school you have completed or the highest degree you have received?}
\item{top_rev}{An ideal society requires some groups to be on top and others to be on the bottom}
\item{inferior_rev}{Some groups of people are simply inferior to other groups}
\item{dominate_flip}{One group should dominate in society}
\item{deserving_flip}{Groups at the bottom are not just as deserving as groups at the top}
\item{sdo_sum}{Social Dominance Orientation}
\item{sdo_avg}{Social Dominance Orientation}
\item{pol_part_sum}{Political Participation}
\item{controlled_rev}{Much of our lives are being controlled by plots hatched in secrecy}
\item{small_rev}{Even though we live in a democracy, a few people will always run things anyway}
\item{run_rev}{The people who really "run" the country are not known to the voters}
\item{big_events_rev}{Big events like wars, recessions, and the outcomes of elections are controlled by small groups of people who are working in secret against the rest of us}
\item{acts_sum}{Conspiracy Theory Belief}
\item{acts_avg}{Conspiracy Theory Belief}
}
}
\usage{
test_data
}
\description{
{dataset_description},
}
\keyword{datasets}
