% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_coef}
\alias{theme_coef}
\title{Theme for coefficient plots}
\usage{
theme_coef(
  base_size = 12,
  grid_x_only = TRUE,
  facet_title_margin_top = base_size,
  facet_title_margin_bottom = half_line,
  facet_title_margin_right = 0.8 * half_line,
  facet_title_margin_left = 0.8 * half_line,
  markdown = FALSE,
  ...
)
}
\arguments{
\item{base_size}{Base font size, given in pts. Also controls the spacing in
the graph.}

\item{grid_x_only}{Logical. Determines if only x-axis grid lines (vertical
lines) should appear. If \code{FALSE}, the default, all grid lines appear. If
\code{TRUE}, only the x-axis grid lines appear and the y-axis grid lines will
disappear.}

\item{...}{Other arguments passed onto \code{theme_default()}.}
}
\description{
This function creates a theme for coefficient plots.
}
