% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_default}
\alias{theme_default}
\title{Default theme}
\usage{
theme_default(
  base_size = 12,
  base_family = "L",
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  axis_text = TRUE,
  axis_text_x = TRUE,
  axis_text_y = TRUE,
  grid = TRUE,
  grid_x_only = FALSE,
  grid_y_only = FALSE,
  grid_major = TRUE,
  grid_minor = TRUE,
  grid_major_x = TRUE,
  grid_major_y = TRUE,
  grid_minor_x = TRUE,
  grid_minor_y = TRUE
)
}
\arguments{
\item{base_size}{Base font size, given in pts. Also controls
the spacing in the graph.}

\item{base_family}{Base font family. Default is Lato.}

\item{base_line_size}{Base size for line elements.}

\item{base_rect_size}{Base size for rect elements.}

\item{axis_text}{Logical. Determines if BOTH axes have labels. If \code{TRUE}, the
default, both axes are labelled. If \code{FALSE}, neither axis is labelled.
Note, this controls both axes. If you want to remove only one axis, use the
\code{axis_text_x} or \code{axis_text_y}.}

\item{axis_text_x}{Logical. Determines if the x-axis has labels. If \code{TRUE},
the default, the x-axis labels are shown. If \code{FALSE}, the x-axis labels are
removed from the plot.}

\item{axis_text_y}{Logical. Determines if the y-axis has labels. If \code{TRUE},
the default, the y-axis labels are shown. If \code{FALSE}, the y-axis labels are
removed from the plot.}

\item{grid}{Logical. Determines if ALL grid lines should appear. If \code{TRUE},
the default, all grid lines appear. If \code{FALSE}, all grid lines disappear.}

\item{grid_x_only}{Logical. Determines if only x-axis grid lines (vertical
lines) should appear. If \code{FALSE}, the default, all grid lines appear. If
\code{TRUE}, only the x-axis grid lines appear and the y-axis grid lines will
disappear.}

\item{grid_y_only}{Logical. Determines if only y-axis grid lines (horizontal
lines) should appear. If \code{FALSE}, the default, all grid lines appear. If
\code{TRUE}, only the y-axis grid lines appear and the  x-axis grid lines will
disppear.}

\item{grid_major}{Logical. Determines if the major grid lines should appear.
If \code{TRUE}, the default, the major grid lines will appear. If \code{FALSE}, the
major grid lines will disappear.}

\item{grid_minor}{Logical. Determines if the minor grid lines should appear.
If \code{TRUE}, the default, the minor grid lines will appear. If \code{FALSE}, the
minor grid lines will disappear.}

\item{grid_major_x}{Logical. Determines if the major x-axis grid lines will
appear. If \code{TRUE}, the default, the major x-axis grid lines will appear.
If \code{FALSE}, the major x-axis grid lines will disappear.}

\item{grid_major_y}{Logical. Determines if the major y-axis grid lines will
appear. If \code{TRUE}, the default, the major y-axis grid lines will appear.
If \code{FALSE}, the major t-axis grid lines will disappear.}

\item{grid_minor_x}{Logical. Determines if the minor x-axis grid lines will
appear. If \code{TRUE}, the default, the minor x-axis grid lines will appear.
If \code{FALSE}, the minor x-axis grid lines will disappear.}

\item{grid_minor_y}{Logical. Determines if the minor y-axis grid lines will
appear. If \code{TRUE}, the default, the minor y-axis grid lines will appear.
If \code{FALSE}, the minor t-axis grid lines will disappear.}
}
\description{
This function creates the default theme that all ADL theme functions are
built off of. It functions similarly to \code{\link[ggplot2]{theme_gray}},
in that all of the default ggplot2 themes are built off
\code{\link[ggplot2]{theme_gray}}.
}
\examples{
# This is designed to make it easy to control various aspects of the graph.
# I will demonstrate a variety of different things that this is capable of.

# First let's create the plot object. We are going to use code from Alison
# Horst's palmerpenguins package:

install.packages("palmerpenguins")
library(palmerpenguins)
library(ggplot2)
library(adlgraphs)

mass_flipper <- ggplot(data = penguins,
  aes(x = flipper_length_mm,
      y = body_mass_g)) +
  geom_point(aes(color = species,
                 shape = species),
             size = 3,
             alpha = 0.8) +
  scale_color_manual(values = c("darkorange", "purple", "cyan4")) +
  labs(title = "Penguin size, Palmer Station LTER",
       subtitle = "Flipper length and body mass for Adelie, Chinstrap and Gentoo Penguins",
       x = "Flipper length (mm)",
       y = "Body mass (g)",
       color = "Penguin species",
       shape = "Penguin species")

# The basic graph with no customizations:
mass_flipper + theme_default()


}
