% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{write_word_table}
\alias{write_word_table}
\title{Internal function to be used when we export to a word doc}
\usage{
write_word_table(var, doc)
}
\description{
Internal function to be used when we export to a word doc
}
