% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_sd.R
\name{wtd_sd}
\alias{wtd_sd}
\title{Calculate weighted standard deviation}
\usage{
wtd_sd(x, wt = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector that you want to calculate the SD}

\item{wt}{A numeric vector indicating the weights used in the calculation}

\item{na.rm}{Logical. Indicates if NAs should be removed or not}
}
\value{
A number indicating the weighted SD of \code{x}
}
\description{
While this is mainly an internal function, it is available for everyone
since I could not find a function that calculates a weighted standard
deviation.
}
\details{
Thank you to the team at Hmisc for the creation of \code{\link[Hmisc:wtd.stats]{Hmisc::wtd.var()}}.
This function could not have been created without it. I simplified their
source code and then took the square root.
}
