% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlgraphs-package.R
\docType{package}
\name{adlgraphs-package}
\alias{adlgraphs}
\alias{adlgraphs-package}
\title{adlgraphs: Create Graphs with ADL Style and Branding}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of adlgraphs is to to help ensure that the data visualizations produced by the Center for Antisemitism Research (CAR) at the Anti-Defamation League (ADL) are consistent with ADL's brand guidelines and that CAR's data visualizations are consistent and cohesive. In addition, it contains a myriad array of different functions to help make data analysis and data cleaning quicker, easier, and more succinct.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JDenn0514/adlgraphs}
  \item \url{https://jdenn0514.github.io/adlgraphs/}
  \item Report bugs at \url{https://github.com/JDenn0514/adlgraphs/issues}
}

}
\author{
\strong{Maintainer}: Jacob Dennen \email{jdennen@adl.org}

}
\keyword{internal}
