% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlgraphs-package.R
\docType{package}
\name{adlgraphs-package}
\alias{adlgraphs}
\alias{adlgraphs-package}
\title{adlgraphs: Create Graphs with ADL Style and Branding}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The main goal of adlgraphs is toward improving the efficiency and ease of cleaning, analyzing, and visualizing survey data, so that the turnaround time between data collection and presentation of analysis and findings could be as minimized as much as possible. In addition, this has ensured that the presentation of data and results is consistent across projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JDenn0514/adlgraphs}
  \item \url{https://jdenn0514.github.io/adlgraphs/}
  \item Report bugs at \url{https://github.com/JDenn0514/adlgraphs/issues}
}

}
\author{
\strong{Maintainer}: Jacob Dennen \email{jdennen@adl.org}

}
\keyword{internal}
