% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_var_label.R
\name{attr_var_label}
\alias{attr_var_label}
\title{Get variable label}
\usage{
attr_var_label(x, df, unlist)
}
\arguments{
\item{x}{A vector object, the name of a column in a \code{data.frame}, or an
an actual \code{data.frame} object.}

\item{df}{A \code{data.frame} or \code{tibble} object. This should only be specified
when \code{x} is only the name of a column in a \code{data.frame}.}

\item{unlist}{Logical. If \code{TRUE}, the default, returns a named vector. If
\code{FALSE}, returns a list. This only works when \code{x} is a \code{data.frame}}
}
\description{
This function makes it easy to get the variable labels from either an
individual vector or a data frame. NOTE: it is not possible to set or modify
the variable labels with this function.
}
