% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky_freqs.R
\name{funky_freqs}
\alias{funky_freqs}
\title{Calculate weighted frequencies}
\usage{
funky_freqs(data, x, group, wt, drop_zero = FALSE, decimals = 3, na.rm = TRUE)
}
\arguments{
\item{data}{An object of type data.frame or tibble. If piping the data into the
function, this is not required.}

\item{x}{Either a character string or symbol. The variable with which want
to get the frequencies.}

\item{group}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> A selection of columns to
group the data by in addition to \code{treats}. This operates very similarly
to \code{.by} from dplyr (for more info on that see \link[dplyr:dplyr_by]{?dplyr_by}).}

\item{wt}{Weights. Add if you have a weighting variable and want to get
weighted frequencies.}

\item{drop_zero}{Logical. Determines if rows with 0 should be removed
Default is \code{FALSE}.}

\item{decimals}{Number of decimals each number should be rounded to. Default
is 3.}

\item{na.rm}{Logical. Determines if NAs should be kept or removed Default is
\code{TRUE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use this function to calculate simple weighted frequencies.
You can also specify a grouping variable by which you want to calculate the
frequencies.

The \code{x}, \code{group}, and \code{wt} arguments can either be strings or symbols
(meaning they can have quotes or no quotes). The benefit of this is that it
makes it really easy to iterate this function over a list or vector of
variables with other functions like \code{\link[purrr:map]{purrr::map()}} \code{\link[purrr:map]{purrr::map()}} or \code{\link[purrr:map]{purrr::walk()}}
\code{\link[purrr:map]{purrr::walk()}} that are found in the \code{purrr} package.
}
