% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{get_all_freqs}
\alias{get_all_freqs}
\title{Export frequencies for a set of variables to a word doc.}
\usage{
get_all_freqs(df, var, group1, group2, wt, file_name)
}
\arguments{
\item{df}{An object of type data.frame or tibble. If piping the df into the
function, this is not required.}

\item{var}{A vector of variables you want to get the frequencies for.}

\item{group1}{A character string. The first grouping variable.}

\item{group2}{A character string. The second grouping variable.}

\item{wt}{A character string. Add if you have a weighting variable and want
to get weighted frequencies}
}
\description{
This function uses \code{freq_fun} to get the frequencies for a set of variables
suppplied by the user. It then outputs these frequencies to a word doc.
}
