% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_dupe.R
\name{get_bot_dupe}
\alias{get_bot_dupe}
\title{Get the bots or duplicates from a data frame}
\usage{
get_bot_dupe(data)
}
\arguments{
\item{data}{A data.frame object you want to operate on}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{get_bot_dupe()} is deprecated because the function \code{get_bogus()}
does the same thing but also removes speedsters.

This function is the opposite of \code{remove_bot_dupes()}. Instead of removing bots
and duplicates from a data frame, it keeps them. This was designed to make it easy
to send to survey panel providers so they can remove them from your data set.
}
