% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{get_freq_table}
\alias{get_freq_table}
\title{Get the frequencies as a GT table}
\usage{
get_freq_table(df, var, group, wt)
}
\arguments{
\item{df}{An object of type data.frame or tibble. If piping the df into the
function, this is not required.}

\item{var}{Either a character string or symbol. The variable with which want
to get the frequencies.}

\item{group}{Either a character string or a symbol. The grouping variable.}

\item{wt}{Weights. Add if you have a weighting variable and want to get
weighted frequencies}
}
\description{
This function creates a GT table of the frequencies of a specified variable
and has the ability to get the frequencies for one grouping variable.
While this can be used on it's own, it was created more to be used in
\code{\link[=get_all_freqs]{get_all_freqs()}}`, a function that outputs the frequencies of a set of
variables to a word doc.
}
