% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dicho.R
\name{make_dicho}
\alias{make_dicho}
\title{Make dichotomous factors}
\usage{
make_dicho(df = data, var)
}
\arguments{
\item{df}{The name of the data frame. Usually you'll be piping in the data
so you should put \code{.}. However, if used inside of dplyr::across() then you
need to use an anonymous function like this \verb{\\(var) make_dicho(., var)}.}

\item{var}{A vector or variable you're transforming.}
}
\description{
Create a vector of a dichotomous factor. Currently works for variables with
"Mostly true" through "Mostly false" and "Strongly Agree" through "Strongly
disagree". Use inside of dplyr::mutate() or dplyr::across() to create a new
column of the vectors
}
