% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_rev.R
\name{num_rev}
\alias{num_rev}
\title{Reverse a numeric function}
\usage{
num_rev(x)
}
\arguments{
\item{x}{A vector of class \code{haven_labelled} or \code{numeric}}
}
\description{
Reverse a numeric variable while maintaining variable and value labels
if available. Also adds an attribute describing the transformation the
original variable underwent.
}
\examples{

library(dplyr)
library(labelled)
library(adlgraphs)
# create fake data
df <- tibble::tibble(
  w = sample(1:4, 10, replace = TRUE),
  x = sample(1:4, 10, replace = TRUE),
  y = sample(1:4, 10, replace = TRUE),
  z = sample(1:4, 10, replace = TRUE)
) \%>\%
  # add value labels to x and y but not z to see what happens with no labels
  labelled::set_value_labels(
    x = c(`Strongly agree` = 1,
          `Somewhat agree` = 2,
          `Somewhat disagree` = 3,
          `Strongly disagree` = 4),
    y = c(`Strongly agree` = 1,
          `Somewhat agree` = 2,
          `Somewhat disagree` = 3,
          `Strongly disagree` = 4),
  ) \%>\%
  # set variable labels to x and z
  labelled::set_variable_labels(
    x = "This is the variable label for x",
    z = "This is the variable label for z"
  )


rev_df <- df \%>\% mutate(rev_w = num_rev(w))

head(rev_df)

str(rev_df)

}
