% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_conv.R
\name{pct_conv}
\alias{pct_conv}
\title{Make proper percent labels}
\usage{
pct_conv(data, x = pct, decimals = 1)
}
\arguments{
\item{data}{A data frame or vector. Can be left blank if used during piping}

\item{x}{a variable we want to convert to a percentage. The value is \code{pct}
by default and should always be pct.}

\item{decimals}{Number of decimal places the percent should be rounded to}
}
\value{
The original data.frame found in \code{data} with two changes. The column
called "pct" is multiplied by 100 and a new column is created called "pct_lab"
that is the same as pct but with a "\%" symbol at the end of it.
}
\description{
This function is a wrapper around a very common data transformation that is
done every time we make a frequency plot. It changes the \code{x} variable by
multiplying it 100 and creates a new variable called \code{pct_lab} which is a
string of the \code{x} variable but with a "\%" symbol added.
}
\examples{
library(dplyr)
# get the frequencies of top andupdate 
test_data \%>\%
  get_freqs(top) \%>\% 
  pct_conv()

}
