% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_conv.R
\name{pct_conv}
\alias{pct_conv}
\title{Make proper percent labels}
\usage{
pct_conv(data, x = pct, digits = 1)
}
\arguments{
\item{data}{A data frame or vector. Can be left blank if used during piping}

\item{x}{a variable we want to convert to a percentage. The value is \code{pct}
by default and should always be pct.}

\item{digits}{Number of decimal places the percent should be rounded to}
}
\description{
This function is a wrapper around a very common data transformation that is
done every time we make a frequency plot
}
\examples{

library(tibble)
library(dplyr)
library(labelled)
library(haven)

# create the fake data
df <- tibble::tribble(
  ~x, ~y, ~z,
  3, 2, 3,
  4, 4, 2,
  2, 6, 1,
  1, 1, 4,
  5, 4, 3,
  6, 5, 6
)

df \%>\%
  count(x) \%>\%
  mutate(pct = prop.table(n)) \%>\%
  pct_conv()


}
