% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettytable.R
\name{prettytable}
\alias{prettytable}
\title{Make pretty HTML tables}
\usage{
prettytable(x)
}
\arguments{
\item{x}{An object to turn into a pretty table.}
}
\description{
Takes a data frame object and makes it into a pretty table. This is designed
to work within the \code{adlgraphs} package but it should work with most objects
of class \code{tbl_df}, \code{tbl}, or \code{data.frame}. Very important note, at the moment,
this only works with frequency tables that have columns labelled \code{pct} and
\code{n}. I do not see this changing for frequency tables. Nevertheless, I am
working on adding in more functionality, including for mean tables, tables of
factor loadings, linear regression coefficients, and others.
}
