% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\title{Assign or reorder factor levels manually}
\usage{
refactor(x, new_levels, ordered = NA)
}
\arguments{
\item{x}{A factor or character vector}

\item{new_levels}{A list of the new levels of the factor}

\item{ordered}{Logical. Specifies if the factor is ordered. Default is NA,
which checks to see if the factor is ordered and then uses that to determine
if it should be ordered}
}
\value{
A factor variable of the same length as \code{x}
}
\description{
This is a low level function that allows you to convert a character vector
into a factor and manually assign the levels, or to manually reassign the
levels of a factor vector. While very similar to \code{factor()} a key difference
is that this keeps the original attributes of \code{x}.
}
\examples{
# load the dplyr library so we can use `mutate()`
library(dplyr)

# let's manually reorder the factor levels of `edu_f` from the `test_data`
# data set so it's in a random order that I specify
test_data <- test_data \%>\%
  mutate(
    # make the new reordered variable
    edu_f_reordered = refactor(
      # specify we are reordering the `edu_f` variable
      x = edu_f,
      new_levels = c(
        "Bachelor's Degree",
        "Graduate Degree",
        "High School or Less",
        "Some College"
      )
    )
  )

# let's check the new levels
levels(test_data$edu_f_reordered)
# and compare them to the original levels
levels(test_data$edu_f)
}
