% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_quant.R
\name{stat_density_quant}
\alias{stat_density_quant}
\title{Stat for density ridgeline plots}
\usage{
stat_density_quant(
  mapping = NULL,
  data = NULL,
  geom = geom,
  position = "stack",
  ...,
  bw = "sj",
  adjust = 1,
  kernel = "gaussian",
  n = 512,
  bounds = c(-Inf, Inf),
  show.legend = NA,
  inherit.aes = TRUE,
  quantile_lines = FALSE,
  calc_ecdf = FALSE,
  quantiles = 4,
  quantile_fun = quantile
)
}
\arguments{
\item{geom}{The geometric object to use to display the data. \code{\link{geom_density_quant}}
is the default.}

\item{bw}{The smoothing bandwidth to be used.
If numeric, the standard deviation of the smoothing kernel.
If character, a rule to choose the bandwidth, as listed in
\code{\link[stats:bandwidth]{stats::bw.nrd()}}. Note that automatic calculation of the bandwidth does
not take weights into account. Default is \code{sj}.}

\item{adjust}{A multiplicate bandwidth adjustment. This makes it possible
to adjust the bandwidth while still using the a bandwidth estimator.
For example, \code{adjust = 1/2} means use half of the default bandwidth.}

\item{kernel}{Kernel. See list of available kernels in \code{\link[=density]{density()}}.}

\item{n}{The number of equally spaced points at which the density is to be
estimated. Should be a power of 2. Default is 512.}

\item{bounds}{Known lower and upper bounds for estimated data. Default
\code{c(-Inf, Inf)} means that there are no (finite) bounds. If any bound is
finite, boundary effect of default density estimation will be corrected by
reflecting tails outside \code{bounds} around their closest edge. Data points
outside of bounds are removed with a warning.}

\item{calc_ecdf}{If \code{TRUE}, \code{stat_density_ridges} calculates an empirical
cumulative distribution function (ecdf) and returns a variable \code{ecdf} and
a variable \code{quantile}. Both can be mapped onto aesthetics via \code{stat(ecdf)}
and \code{stat(quantile)}, respectively.}

\item{quantiles}{Sets the number of quantiles the data should be broken into.
Used if either \code{calc_ecdf = TRUE} or \code{quantile_lines = TRUE}. If
\code{quantiles} is an integer then the data will be cut into that many equal
quantiles. If it is a vector of probabilities then the data will cut by them.}

\item{quantile_fun}{Function that calculates quantiles. The function needs
to accept two parameters, a vector \code{x} holding the raw data values and a
vector \code{probs} providing the probabilities that define the quantiles.
Default is \code{quantile}.}

\item{from, to}{The left and right-most points of the grid at which the density is to be estimated,
as in \code{\link[=density]{density()}}. If not provided, these are estimated from the data range and the bandwidth.
and sets it to \code{TRUE}.}
}
\description{
This stat is the default stat used by \code{\link{geom_density_quant}}. It is very
similar to \code{\link[ggplot2]{stat_density}} and
\code{\link[ggridges]{stat_density_ridges}} as it was built as a sort of
combination of the two. One of the key differences between this function and
those two is that this one uses the Sheather & Jones ("sj") as the default bandwidth
selector. This is done because this is a better bandwidth selector than
Silverman's ("nrd0") which is the default for the other two functions. In
addition, this function allows you to add quantile lines similar to
\code{\link[ggridges]{stat_density_ridges}}.
}
