% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{write_word_table}
\alias{write_word_table}
\title{This mainly an internal package but can be used externally}
\usage{
write_word_table(x, doc)
}
\description{
This mainly an internal package but can be used externally
}
\keyword{internal}
